+++ tests/unit/components/Cancel.spec.ts
@@ -1,6 +1,6 @@
 import { shallowMount, createLocalVue } from "@vue/test-utils";
 import Vuex from "vuex";
-import Login from "../../src/pages/Login.vue";
+import Cancel from "@/pages/cencel/Cancel.vue";
 import { QiitaModule } from "@/store/modules/qiita";
 
 import { IQiitaState } from "@/types/qiita";
@@ -13,18 +13,18 @@ localVue.use(VueRouter);
 
 const router = new VueRouter();
 
-describe("Login.vue", () => {
+describe("Cancel.vue", () => {
   let store: any;
   let state: IQiitaState;
   let actions: any;
 
-  it('calls store action "login" when button is clicked', () => {
+  it('calls store action "cancel" when button is clicked', () => {
     state = {
       authorizationCode: "",
       accessToken: "",
       qiitaAccountId: "",
       permanentId: "",
-      sessionId: "",
+      sessionId: "d690e4de-0a4e-4f14-a5c5-f4303fbd8a08",
       categories: [],
       stocks: [],
       categorizedStocks: [],
@@ -37,7 +37,7 @@ describe("Login.vue", () => {
     };
 
     actions = {
-      login: jest.fn()
+      cancel: jest.fn()
     };
 
     store = new Vuex.Store({
@@ -51,10 +51,10 @@ describe("Login.vue", () => {
       }
     });
 
-    const wrapper = shallowMount(Login, { store, localVue, router });
+    const wrapper = shallowMount(Cancel, { store, localVue, router });
     const button = wrapper.find("button");
 
     button.trigger("click");
-    expect(actions.login).toHaveBeenCalled();
+    expect(actions.cancel).toHaveBeenCalled();
   });
 });
