+++ tests/unit/components/CategorizedStock.spec.ts
