+++ tests/unit/components/CategorizedStockEdit.spec.ts
