+++ tests/unit/components/CategorizedStockList.spec.ts
