+++ tests/unit/components/CategoryList.spec.ts
