+++ tests/unit/components/SideMenu.spec.ts
