+++ tests/unit/components/StockEdit.spec.ts
