+++ tests/unit/components/StockList.spec.ts
