+++ tests/unit/components/Stocks.spec.ts
