+++ tests/unit/pages/Error.spec.ts
@@ -7,12 +7,12 @@ describe("Cancel.vue", () => {
   };
 
   it("should have the correct props", () => {
-    const wrapper = shallowMount(Error, { propsData, stubs: ['router-link']});
+    const wrapper = shallowMount(Error, { propsData, stubs: ["router-link"] });
     expect(wrapper.props()).toEqual(propsData);
   });
 
   it("renders the correct error message", () => {
-    const wrapper = shallowMount(Error, { propsData, stubs: ['router-link'] });
+    const wrapper = shallowMount(Error, { propsData, stubs: ["router-link"] });
     expect(wrapper.find("h2").text()).toBe(propsData.errorMessage);
   });
 });
