+++ tests/unit/pages/Stock.spec.ts
