+++ tests/unit/pages/StockCategories.spec.ts
