+++ tests/unit/store/QiitaModule.spec.ts
